corrections

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include <stdio.h>
 /**
  * print_list - prints all the elements of a list_t
  * @h: pointer to list
@@ -7,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int k = 0;
+	size_t k = 0;
 
 	while (h)
 	{
